Use the correct type for pulse train lengths
Pulse train lengths are defined as uint8_t in the protocol datatype, so
they cannot exceed its range anyway. This unifies the propagation
trough the library to reflect this fact clearly.

NOTE: Users of the PulseTrainCallBack feature need to update their
implementation as the signature of the callback has cahnged. The length
is now propagated as size_t. See at the changes of
tests/test_parse/test_parse.ino in this commit as an example.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>