ESPiLight.cpp: replace pilight json generation.
The pilight jdon generation has the problem that it has a buffer that
rets reallocated with twice the size every time it is too small. This
causes the ESP to crash if we're <10k of free memory in some cases.

So this replaces this by a simple hand-written json generation that
uses in advance how much space it needs.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>